% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{create_decompressor}
\alias{create_decompressor}
\title{Create a new decompressor object}
\usage{
create_decompressor(wbits = 0L)
}
\arguments{
\item{wbits}{The window size bits parameter. Default is 0.}
}
\value{
A SEXP pointer to the new decompressor object.
}
\description{
Initialize a new decompressor object for zlib-based decompression.
}
\examples{
decompressor <- create_decompressor()
}
