% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{decompress_chunk}
\alias{decompress_chunk}
\title{Decompress a chunk of data}
\usage{
decompress_chunk(decompressorPtr, input_chunk)
}
\arguments{
\item{decompressorPtr}{An external pointer to an initialized decompressor object.}

\item{input_chunk}{A raw vector containing the compressed data chunk.}
}
\value{
A raw vector containing the decompressed data.
}
\description{
Perform chunk-wise decompression on a given raw vector using a decompressor object.
}
\examples{
rawToChar(decompress_chunk(create_decompressor(), memCompress(charToRaw("Hello, World"))))
}
