% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quiet.R
\name{quiet}
\alias{quiet}
\title{Force a function to not print anything}
\usage{
quiet(x)
}
\arguments{
\item{x}{A function invocation.}
}
\value{
whatever the function returns, but invisible.
}
\description{
Force a function to not print anything
}
\examples{
quiet(cat(5))
quiet((function(x) {cat(x); x^2})(5))
}
