% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_renv_bib.R
\name{write_renv_bib}
\alias{write_renv_bib}
\title{Create a bibliography file of all renv package dependencies}
\usage{
write_renv_bib(filepath = "./pkgs.bib")
}
\arguments{
\item{filepath}{The ('.bib') file to write, passed on to knitr::write_bib().}
}
\value{
(invisibly) a character vector of packages which was passed on to
    knitr::write_bib().
}
\description{
See \url{https://github.com/rstudio/renv/issues/340} for more detail
}
