% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compress}
\alias{compress}
\title{Compress a raw vector using zlib}
\usage{
compress(x, level = -1L)
}
\arguments{
\item{x}{a raw vector to be compressed}

\item{level}{the compression level as an integer from -1 to 9, where
1 means faster processing and less compression, 9 means the highest
compression but slowest processing. 0 perfoms no compression and -1
is the default which is a balance of speed and compression 
(currently equivalent to 6).}
}
\description{
Compress a raw vector using zlib
}
