% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-themes.R
\name{theme_ms}
\alias{theme_ms}
\title{Basic plot theme for manuscript}
\usage{
theme_ms(
  base_size = 12,
  text_size_strip = 16,
  text_size_axis_text = 12,
  text_size_axis_title = 18,
  text_size_title = 20,
  text_size_subtitle = 16,
  text_size_legend = 16,
  text_size_legend_title = 18,
  text_size_caption = 14,
  font_family = "",
  legend_position = "bottom",
  panel_spacing_lines = 2,
  margin_t = 6,
  margin_r = 6,
  margin_b = 6,
  margin_l = 6,
  ...
)
}
\arguments{
\item{base_size}{base font size}

\item{text_size_strip}{size of text that goes over facets}

\item{text_size_axis_text}{size of axis tick text}

\item{text_size_axis_title}{size of axis label/title text}

\item{text_size_title}{size of plot title text}

\item{text_size_subtitle}{size of plot subtitle text}

\item{text_size_legend}{size of legend text}

\item{text_size_legend_title}{size of legend title text}

\item{text_size_caption}{size of plot caption text}

\item{font_family}{font to use in plot text, defaults to ggplot's base font}

\item{legend_position}{where the legend should go, bottom by default.
accepts any argument allowed by ggplot2.}

\item{panel_spacing_lines}{Amount of space in line units between faceted
panels.}

\item{margin_t}{Margin space above top of plot.}

\item{margin_r}{Margin space to the right of plot.}

\item{margin_b}{Margin space underneath the of plot.}

\item{margin_l}{Margin space to the left of plot.}

\item{...}{Other arguments passed to \code{theme_minimal}.}
}
\description{
Basic plot theme for manuscript
}
\examples{
\dontrun{
library(ggplot2)

ggplot(iris, aes(Sepal.Length, Petal.Length)) +
    geom_point() +
    theme_ms()
}

\dontrun{
ggplot2::theme_set(zlib::theme_ms())
}

}
